refactor: move state flags and parsing functions into state pkg
