+++ cmd/celestia/bridge.go
@@ -5,6 +5,7 @@ import (
 
 	cmdnode "github.com/celestiaorg/celestia-node/cmd"
 	"github.com/celestiaorg/celestia-node/node/node"
+	"github.com/celestiaorg/celestia-node/node/state"
 )
 
 // NOTE: We should always ensure that the added Flags below are parsed somewhere, like in the PersistentPreRun func on
@@ -18,7 +19,7 @@ func init() {
 			cmdnode.CoreFlags(),
 			cmdnode.MiscFlags(),
 			cmdnode.RPCFlags(),
-			cmdnode.KeyFlags(),
+			state.KeyFlags(),
 		),
 		cmdnode.Start(
 			cmdnode.NodeFlags(node.Bridge),
@@ -26,7 +27,7 @@ func init() {
 			cmdnode.CoreFlags(),
 			cmdnode.MiscFlags(),
 			cmdnode.RPCFlags(),
-			cmdnode.KeyFlags(),
+			state.KeyFlags(),
 		),
 	)
 }
@@ -68,7 +69,10 @@ var bridgeCmd = &cobra.Command{
 			return err
 		}
 
-		ctx = cmdnode.ParseKeyFlags(ctx, cmd)
+		opts := state.ParseKeyFlags(cmd)
+		if len(opts) > 0 {
+			ctx = cmdnode.WithNodeOptions(ctx, opts...)
+		}
 
 		cmd.SetContext(ctx)
 		return nil
