+++ node/state/opts.go
@@ -1,6 +1,8 @@
 package state
 
-import "github.com/celestiaorg/celestia-node/node/node"
+import (
+	"github.com/celestiaorg/celestia-node/node/node"
+)
 
 // WithKeyringAccName sets the `KeyringAccName` field in the key config.
 func WithKeyringAccName(name string) node.Option {
